All Major Components Working as of Commit
Code is running in a stable manner.
Piping from one process to another is working.
Redirection of output to a file is working.
Redirection of input from a file is working.

Examples of commands to run:
cat < ‘file_name’’ | grep ’selector’ > ‘file name’ — Cat will output
the file given into the pipe, grep will then filter the input from the
pipe based on the selector and output the results to a file called
‘file_name’ (will over write the file if it exists, if not, it will be
created).

The change directory builtin command is working also.

Adam.